Implement hypercube and mesh oriented DOR routing.
This is dependent on the switches being consistently cabled in
port-to-dimension order.  It is activated by setting the Hypercube tag
to 1 in the sm section of the xml file.  Either the shortestpath or
dgshortestpath RoutingAlgorithm can be used.  Traffic is distributed
only across the port group associated with the lowest dimension to a
neighbor switch along a shortest path.  The cost of a hop needs to be
constant so that it follows dimension ordering rules and is not
influenced by differing port speeds.