Implement a switch port order and cost feature.
This affects the order that ports are examined for the hypercube
routing.  This effectively provides an override to the port <->
dimension mapping.  It can also affect the distribution of traffic to
ports for (dg)shortestpath routing.  The cost of each port can also
be altered, which will affect the shortest path calculation.