Fix for nodes dropping off from IPoIB mcast group
Check if the old topology port was DOWN when determining if
ClientReregister should be sent to handle the case where the
FM marked a port DOWN on an error path.

This is needed because if the port was marked DOWN, then we
will have cleaned up the MC groups.

Update version in spec file as well as GitHub url

Signed-off-by: John Fleck <john.fleck@intel.com>
Signed-off-by: Michael Brooks <michael.brooks@intel.com>