Incorporate use LIBDIR patch
Pull in and adjust patch from:
https://github.com/cornelisnetworks/opa-fm/pull/33
by Benson Muite.

We should allow the rpmmacro to set the LIBDIR but all of our documentation
and scripts assume /usr/lib not /usr/lib64. We need to force /usr/lib for
now so updated spec file to gurantee this.

These RPM Build warnings are a result of not allowing the LIBDIR macro to
dictate things I believe.

RPM build warnings:
    absolute symlink: /usr/sbin/opafmcmd -> /usr/lib/opa-fm/bin/fm_cmd
    absolute symlink: /usr/sbin/opafmcmdall -> /usr/lib/opa-fm/bin/fm_cmdall
    absolute symlink: /usr/sbin/opafmconfigpp -> /usr/lib/opa-fm/bin/opafmconfigpp
    absolute symlink: /usr/sbin/opafmvf -> /usr/lib/opa-fm/bin/opafmvf

Signed-off-by: Dennis Dalessandro <dennis.dalessandro@cornelisnetworks.com>